+++ VisualTests/Tests/MSAA.cpp
@@ -36,7 +36,7 @@ namespace VisualTests {
         imageMSAA = info.resourceManager.CreatePersistentImage({
             .format = Format::RGBA8Unorm,
             .size = { info.displayInfo.width, info.displayInfo.height },
-            .sampleCount = static_cast<u32>(sampleCount),
+            .sampleCount = sampleCount,
             .usage = ImageUsageFlagBits::RENDER_TARGET,
             .name = "MSAA Render Image",
         });
